- added VideoSection to first session dialog
